move snapstate functions into snapstate.go
