+++ overlord/snapstate/snapmgr.go
@@ -25,91 +25,10 @@ import (
 
 	"gopkg.in/tomb.v2"
 
-	"github.com/ubuntu-core/snappy/i18n"
 	"github.com/ubuntu-core/snappy/overlord/state"
 	"github.com/ubuntu-core/snappy/progress"
 	"github.com/ubuntu-core/snappy/snappy"
 )
-
-// Install returns a set of tasks for installing snap.
-// Note that the state must be locked by the caller.
-func Install(s *state.State, snap, channel string, flags snappy.InstallFlags) (*state.TaskSet, error) {
-	t := s.NewTask("install-snap", fmt.Sprintf(i18n.G("Installing %q"), snap))
-	t.Set("state", installState{
-		Name:    snap,
-		Channel: channel,
-		Flags:   flags,
-	})
-
-	return state.NewTaskSet(t), nil
-}
-
-// Update initiates a change updating a snap.
-// Note that the state must be locked by the caller.
-func Update(s *state.State, snap, channel string, flags snappy.InstallFlags) (*state.TaskSet, error) {
-	t := s.NewTask("update-snap", fmt.Sprintf(i18n.G("Updating %q"), snap))
-	t.Set("state", installState{
-		Name:    snap,
-		Channel: channel,
-		Flags:   flags,
-	})
-
-	return state.NewTaskSet(t), nil
-}
-
-// Remove returns a set of tasks for removing snap.
-// Note that the state must be locked by the caller.
-func Remove(s *state.State, snap string, flags snappy.RemoveFlags) (*state.TaskSet, error) {
-	t := s.NewTask("remove-snap", fmt.Sprintf(i18n.G("Removing %q"), snap))
-	t.Set("state", removeState{
-		Name:  snap,
-		Flags: flags,
-	})
-
-	return state.NewTaskSet(t), nil
-}
-
-// Purge returns a set of tasks for purging a snap.
-// Note that the state must be locked by the caller.
-func Purge(s *state.State, snap string, flags snappy.PurgeFlags) (*state.TaskSet, error) {
-	t := s.NewTask("purge-snap", fmt.Sprintf(i18n.G("Purging %q"), snap))
-	t.Set("state", purgeState{
-		Name:  snap,
-		Flags: flags,
-	})
-
-	return state.NewTaskSet(t), nil
-}
-
-// Rollback returns a set of tasks for rolling back a snap.
-// Note that the state must be locked by the caller.
-func Rollback(s *state.State, snap, ver string) (*state.TaskSet, error) {
-	t := s.NewTask("rollback-snap", fmt.Sprintf(i18n.G("Rolling back %q"), snap))
-	t.Set("state", rollbackState{
-		Name:    snap,
-		Version: ver,
-	})
-
-	return state.NewTaskSet(t), nil
-}
-
-// SetActive returns a set of tasks for setting a snap active.
-// Note that the state must be locked by the caller.
-func SetActive(s *state.State, snap string, active bool) (*state.TaskSet, error) {
-	var msg string
-	if active {
-		msg = fmt.Sprintf(i18n.G("Set active %q"), snap)
-	} else {
-		msg = fmt.Sprintf(i18n.G("Set inactive %q"), snap)
-	}
-	t := s.NewTask("set-active-snap", msg)
-	t.Set("state", setActiveState{
-		Name:   snap,
-		Active: active,
-	})
-
-	return state.NewTaskSet(t), nil
-}
 
 // SnapManager is responsible for the installation and removal of snaps.
 type SnapManager struct {
